Search form choices and earch page

+++ listings/models.py
@@ -10,7 +10,7 @@ class Listing(models.Model):
     state = models.CharField(max_length=50)
     zipcode = models.CharField(max_length=20)
     description = models.TextField(blank=True)
-    price = models.FloatField()
+    price = models.IntegerField()
     bedrooms = models.IntegerField()
     bathrooms = models.DecimalField(max_digits=2, decimal_places=1)
     garage = models.IntegerField(default=0)
